+++ tests/test_option_groups.py
@@ -47,13 +47,13 @@ def test_option_groups_are_correctly_displayed_in_help(
 def test_option_group_constraints_are_checked(runner, get_example_command):
     cmd = get_example_command(align_option_groups=False)
 
-    result = runner.invoke(cmd, args='--one=1')
+    result = runner.invoke(cmd, args='arg1 --one=1')
     assert result.exit_code == 0
 
-    result = runner.invoke(cmd, args='--one=1 --three=3 --five=4')
+    result = runner.invoke(cmd, args='arg1 --one=1 --three=3 --five=4')
     assert result.exit_code == 0
 
-    result = runner.invoke(cmd, args='--one=1 --three=3')
+    result = runner.invoke(cmd, args='arg1 --one=1 --three=3')
     assert result.exit_code == 2
     error_prefix = ('Error: when --three is set, at least 1 of the following '
                     'parameters must be set')